feat: Header, Profile 컴포넌트 분리 및 Style 파일 분리
